requirements.txt
requests
tldextract
python-whois
favicon
dnspython
xmltodict
tarfile
datetime